% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{fun}
\alias{fun}
\title{Core function of wmy_1e, wmy_3e, & my_2e}
\usage{
fun(df, metric, grouping, df_rr, metric_rr, df_goal, metric_goal, df_3p9,
  metric_3p9, df_6p6, metric_6p6, df_9p3, metric_9p3, weeks_back = 4,
  week_end_dates = FALSE, week_start_dates = TRUE, cbr_ytd = FALSE,
  full_yr = FALSE, rate = FALSE, new_name = NULL, accounting = FALSE,
  div_by_1000 = FALSE, prefix = "", suffix = "", spark = FALSE,
  pop = FALSE, pop_threshold = 100, scalar = 1, digitsAfterDecimal = 2,
  op2_and_var_ph = TRUE, month_names = TRUE, in_mth_header_op2 = TRUE,
  in_mth_header_3p9 = FALSE, in_mth_header_6p6 = FALSE,
  in_mth_header_9p3 = FALSE)
}
\arguments{
\item{df}{data frame containing metric}

\item{metric}{column name from df}

\item{grouping}{time diminesion column name (wk_num_in_yr, mth_num_in_yr, yr_num)}

\item{df_rr}{data frame containing run rate}

\item{metric_rr}{column name from df_rr}

\item{df_goal}{data frame containing metric goal}

\item{metric_goal}{column name from df_goal}

\item{df_3p9}{data frame containing metric prediction}

\item{metric_3p9}{column name from df_3p9}

\item{df_6p6}{data frame containing metric prediction}

\item{metric_6p6}{column name from df_6p6}

\item{df_9p3}{data frame containing metric prediction}

\item{metric_9p3}{column name from df_9p3}

\item{cbr_ytd}{logical. Inlcude YTD definition for CBR : Actual values from last completed month}

\item{full_yr}{logical. Inlcude full year summation column for month view}

\item{rate}{logical. Adjust logic to handle full year for rates}

\item{new_name}{name for metric output. Also creates new labels for previous year, previous year variance, etc.}

\item{accounting}{format values as accountants do. ie: 1234000 -> 1,234,000 & -3000 -> (3000)}

\item{div_by_1000}{also an accounting practice; divide all value by 1000. ie: 1234000 -> 1234.00}

\item{prefix}{add prefix to values. ie: 2000 -> $2000}

\item{suffix}{add suffix to values. ie: 20.54 -> 20.54\%}

\item{spark}{logical should a spark chart be added as a column in the output}

\item{pop}{logical stands for Period over Period. Include percent change from previous period (ie: week 4 to week 5) next to value}
}
\description{
Core function of wmy_1e, wmy_3e, & my_2e
}
\examples{
## week view
fun(
performance,
revenue,
wk_num_in_yr,
new_name = "Revenue",
accounting = TRUE,
div_by_1000 = FALSE,
full_yr = FALSE,
prefix = "$"
)

## month view
fun(
performance,
revenue,
mth_num_in_yr,
new_name = "Revenue",
accounting = TRUE,
div_by_1000 = FALSE,
full_yr = FALSE,
prefix = "$"
)
}
