% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-performance.R
\docType{data}
\name{performance}
\alias{performance}
\title{Dunder Mifflin KPI data by salesman and day}
\format{A tibble
\describe{
\item{company}{Dunder Mifflin}
\item{branch}{Scranton}
\item{state_abb}{PA}
\item{salesman_full_name}{Jim, Pam, Dwight, etc.}
}}
\usage{
performance
}
\description{
Dunder Mifflin KPI data by salesman and day
}
\examples{
head(performance)
}
\keyword{datasets}
