% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmy_3e.R
\name{wmy_3e}
\alias{wmy_3e}
\title{[w]eek [m]onth [y]ear _ [3] [e]xports}
\usage{
wmy_3e(df_wk, df_mth, df_yr, metric, df_rr, metric_rr, df_goal, metric_goal,
  df_3p9, metric_3p9, df_6p6, metric_6p6, df_9p3, metric_9p3,
  week_end_dates = FALSE, week_start_dates = TRUE, full_yr = TRUE,
  cbr_ytd = TRUE, rate = FALSE, new_name = NULL, commas = FALSE,
  neg_var_paren = TRUE, div_by_1000 = FALSE, prefix = "", suffix = "",
  scalar = 1, digitsAfterDecimal = 2, spark = FALSE, pop = FALSE,
  pop_threshold = 100, in_mth_header_op2 = TRUE,
  in_mth_header_3p9 = FALSE, in_mth_header_6p6 = FALSE,
  in_mth_header_9p3 = FALSE)
}
\arguments{
\item{df_wk}{weekly grain data frame}

\item{df_mth}{monthly grain data frame}

\item{df_yr}{yearly grain data frame}

\item{metric}{column name for metric (uniform across all the grains of data frames)}

\item{df_rr}{data frame containing run rate}

\item{metric_rr}{column name from df_rr}

\item{df_goal}{data frame containing goal}

\item{metric_goal}{column name from df_goal}

\item{df_3p9}{data frame containing 3 plus 9}

\item{metric_3p9}{column name from df_3p9}

\item{df_6p6}{data frame containing 6 plus 6}

\item{metric_6p6}{column name from df_6p6}

\item{df_9p3}{data frame containing 9 plus 3}

\item{metric_9p3}{column name from df_9p3}

\item{week_end_dates}{logical. display week end dates in week portion header}

\item{week_start_dates}{logical. display week start dates in week portion header}

\item{full_yr}{logical. Inlcude full year summation column for month view}

\item{cbr_ytd}{logical. Inlcude YTD definition for CBR : Actual values from last completed month}

\item{rate}{logical. Adjust logic to handle full year for rates}

\item{new_name}{name for metric output. Also creates new labels for previous year, previous year variance, etc.}

\item{commas}{format values with commas. ie: 1234000 -> 1,234,000}

\item{neg_var_paren}{format variance values with parenthesis. ie: -40.4 -> (40.4)}

\item{div_by_1000}{also an accounting practice; divide all value by 1000. ie: 1234000 -> 1234.00}

\item{prefix}{add prefix to values. ie: 2000 -> $2000}

\item{suffix}{add suffix to values. ie: 20.54 -> 20.54\%}

\item{in_mth_header_9p3}{}
}
\description{
[w]eek [m]onth [y]ear _ [3] [e]xports
}
